+++ holiday-calendar-apac/src/test/java/org/holiday/calendar/observance/islamic/apac/HariRayaPuasaTest.java
@@ -16,7 +16,7 @@
  * Temple Place, Suite 330, Boston, MA 02111-1307 USA
  ******************************************************************************/
 
-package org.holiday.calendar.observance.islamic;
+package org.holiday.calendar.observance.islamic.apac;
 
 import org.holiday.calendar.observance.islamic.apac.HariRayaPuasa;
 import org.holiday.calendar.observance.lunar.VesakDay;
